# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: E-Commerce API V1
  version: 1.0.0
  description: |
    Mobil ve admin için REST API. Tüm müşteri endpoint'leri `/api/v1` altında.
    Admin `/api/v1/admin/*`. Locale `?locale=tr` ile.
servers:
  - url: https://api.example.com/api/v1
    description: Production
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Kayıt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Giriş
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Token yenileme
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Çıkış (refresh invalidate)
      responses:
        '204':
          description: No content
...
  /products:
    get:
      tags: [Products]
      summary: Ürün listesi (cursor pagination)
      parameters:
        - $ref: '#/components/parameters/Locale'
        - name: categoryId
          in: query
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
...
  /cart:
    get:
      tags: [Cart]
      summary: Sepet (guest veya user)
      parameters:
        - name: Guest-Token
          in: header
          schema: { type: string }
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    put:
      tags: [Cart]
      summary: Sepet satırlarını güncelle (tam replace veya merge policy backend)
      parameters:
        - name: Guest-Token
          in: header
          schema: { type: string }
      security:
        - bearerAuth: []
        - {}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [variantId, quantity]
                    properties:
                      variantId: { type: string }
                      quantity: { type: integer, minimum: 0 }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /cart/merge:
    post:
      tags: [Cart]
      security: [{ bearerAuth: [] }]
      summary: Misafir sepetini kullanıcıya birleştir
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [guestToken]
              properties:
                guestToken: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /addresses:
    get:
      tags: [Addresses]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags: [Addresses]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /addresses/{id}:
    patch:
      tags: [Addresses]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      tags: [Addresses]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: OK
...
  /orders:
    post:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      summary: Sipariş oluştur + ödeme oturumu (idempotent)
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
                  payment:
                    type: object
                    properties:
                      redirectUrl: { type: string }
                      provider: { type: string }
        '409':
          description: Stok / fiyat değişti

    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  nextCursor: { type: string, nullable: true }
...
  /notifications:
    get:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  nextCursor: { type: string }
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: OK
  /favorites:
    get:
      tags: [Favorites]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSummary'
    post:
      tags: [Favorites]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: string }
      responses:
        '204':
          description: OK
    delete:
      tags: [Favorites]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: productId
          in: query
          required: true
          schema: { type: string }
      responses:
        '204':
          description: OK
...
  /admin/login:
    post:
      tags: [Admin]
      summary: Admin giriş
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  expiresIn: { type: integer }
  /admin/products:
    get:
      tags: [Admin]
      security: [{ adminBearer: [] }]
      responses:
        '200':
          description: Paginated products
    post:
      tags: [Admin]
      security: [{ adminBearer: [] }]
      responses:
        '201':
          description: Created
  /admin/orders:
    get:
      tags: [Admin]
      security: [{ adminBearer: [] }]
      responses:
        '200':
          description: Order list
  /admin/orders/{id}/status:
    patch:
      tags: [Admin]
      security: [{ adminBearer: [] }]
      summary: Sipariş durumu (processing/shipped/delivered)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [processing, shipped, delivered, cancelled]
      responses:
        '200':
          description: OK
  /admin/banners:
    get:
      tags: [Admin]
      security: [{ adminBearer: [] }]
    post:
      tags: [Admin]
      security: [{ adminBearer: [] }]
  /admin/coupons:
    get:
      tags: [Admin]
      security: [{ adminBearer: [] }]
    post:
      tags: [Admin]
      security: [{ adminBearer: [] }]
  /admin/activity-logs:
    get:
      tags: [Admin]
      security: [{ adminBearer: [] }]
      summary: Audit log listesi
      responses:
        '200':
          description: OK
components:
  parameters:
    Locale:
      name: locale
      in: query
      schema:
        type: string
        default: tr
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password:
          type: string
          minLength: 8
        name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
...
    ProductSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        thumbnailUrl: { type: string, nullable: true }
        minPrice: { type: number }
        currency: { type: string }
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
        nextCursor: { type: string, nullable: true }
...
    Cart:
      type: object
      properties:
        guestToken: { type: string, nullable: true }
        items:
          type: array
          items:
            type: object
            properties:
              variantId: { type: string }
              quantity: { type: integer }
              productName: { type: string }
              unitPrice: { type: number }
              lineTotal: { type: number }
        subtotal: { type: number }
        currency: { type: string }
    AddressInput:
      type: object
      required: [title, line1, city, district]
      properties:
        title: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        district: { type: string }
        postalCode: { type: string, nullable: true }
        isDefault: { type: boolean }
    Address:
      allOf:
        - $ref: '#/components/schemas/AddressInput'
        - type: object
          properties:
            id: { type: string }
    CreateOrderRequest:
      type: object
      required: [addressId, couponCode]
      properties:
        addressId: { type: string }
        couponCode: { type: string, nullable: true }
        clientRequestId: { type: string, description: Idempotency body fallback }
    Order:
      type: object
      properties:
        id: { type: string }
        status:
          type: string
          enum:
            [
              pending_payment,
              paid,
              processing,
              shipped,
              delivered,
              cancelled,
            ]
        paymentStatus:
          type: string
        total: { type: number }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
...
    Notification:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        body: { type: string }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }